@@ -36,8 +36,7 @@ dev = [
     "pytest>=8.4.0,<9.0",
     "pytest-cov>=4.0.0",
     "pytest-mock>=3.12.0",
-    "black>=24.0.0",
-    "flake8>=7.0.0",
+    "ruff>=0.8.0,<1.0",
 ]
 test = [
     "pytest>=8.4.0,<9.0",
@@ -60,7 +59,40 @@ where = ["src"]
 [tool.setuptools.package-dir]
 "" = "src"
 
-[tool.black]
+[tool.ruff]
 line-length = 88
-target-version = ['py39']
-include = '\.pyi?$'
+target-version = "py310"
+
+[tool.ruff.lint]
+select = [
+    "E",   # pycodestyle errors
+    "W",   # pycodestyle warnings
+    "F",   # pyflakes
+    "I",   # isort
+    "B",   # flake8-bugbear
+    "C4",  # flake8-comprehensions
+    "UP",  # pyupgrade
+    "N",   # pep8-naming
+    "S",   # flake8-bandit (security)
+    "T20", # flake8-print
+    "SIM", # flake8-simplify
+    "RUF", # Ruff-specific rules
+]
+ignore = [
+    "E501",  # line too long (handled by formatter)
+    "S101",  # use of assert (common in tests)
+    "T201",  # print statements (useful for debugging)
+    "S311",  # random generators (not cryptographic use)
+]
+fixable = ["ALL"]
+unfixable = []
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+[tool.ruff.lint.per-file-ignores]
+"tests/**/*" = ["S101"]  # Allow asserts in tests
+"src/main.py" = ["T201"]  # Allow prints in main entry point
