@@ -1,7 +1,11 @@
-from typing_extensions import TypedDict, Annotated
+from typing import Annotated
+
+from typing_extensions import TypedDict
+
 
 def merge_dicts(a: dict[str, any], b: dict[str, any]) -> dict[str, any]:
     return {**a, **b}
+
 
 class State(TypedDict):
     data: Annotated[dict[str, any], merge_dicts]
