@@ -1,47 +1,41 @@
-import sys
 import argparse
+import sys
 
-from langgraph.graph import StateGraph, START, END
+from langgraph.graph import END, START, StateGraph
 
+from data.models import PortfolioPreference
+from graph.state import State
+from llm.models import get_llm_model, load_models
+from nodes.analyst_agents.analysis_workflow import create_analyst_graph
 from nodes.investment_agents.goal_based import investment_strategy
 from nodes.portfolios_agent import create_portfolio
-from nodes.analyst_agents.analysis_workflow import create_analyst_graph
-
-from graph.state import State
-from utils.display import print_strategy, print_portfolio, print_analysis_response
+from utils.display import print_analysis_response, print_portfolio, print_strategy
 from utils.questionnaires import choose_llm_model, get_user_preferences
-from data.models import PortfolioPreference
-
-from llm.models import load_models, get_llm_model
 
 
 # === Workflow Execution ===
 def run_workflow(
-        show_reasoning:bool, 
-        investment_llm_agent, 
-        portfolio_llm_agent,
-        analyst_llm_agent,
-        analyst:str=None,
-        preferences:PortfolioPreference=None
-    ):
-    
+    show_reasoning: bool,
+    investment_llm_agent,
+    portfolio_llm_agent,
+    analyst_llm_agent,
+    analyst: str = None,
+    preferences: PortfolioPreference = None,
+):
     agent = create_workflow()
 
     result = agent.invoke(
         {
-            'data':{
-                'investment': {
-                    'analyst': analyst,
-                    'user_preferences': preferences
-                },
-                'benchmark': preferences.benchmark
+            "data": {
+                "investment": {"analyst": analyst, "user_preferences": preferences},
+                "benchmark": preferences.benchmark,
             },
-            'metadata': 
-            {'show_reasoning': show_reasoning,
-             'investment_llm_agent': investment_llm_agent,
-             'portfolio_llm_agent': portfolio_llm_agent,
-             'analyst_llm_agent': analyst_llm_agent
-            }
+            "metadata": {
+                "show_reasoning": show_reasoning,
+                "investment_llm_agent": investment_llm_agent,
+                "portfolio_llm_agent": portfolio_llm_agent,
+                "analyst_llm_agent": analyst_llm_agent,
+            },
         }
     )
 
@@ -73,11 +67,18 @@ def create_workflow():
 
     return agent
 
+
 def main():
     """Entry point for the ai-robo-advisor CLI application."""
     try:
-        parser = argparse.ArgumentParser(description="Run the AI Robo Advisor workflow.")
-        parser.add_argument("--show-reasoning", action="store_true", help="Show reasoning from each agent")
+        parser = argparse.ArgumentParser(
+            description="Run the AI Robo Advisor workflow."
+        )
+        parser.add_argument(
+            "--show-reasoning",
+            action="store_true",
+            help="Show reasoning from each agent",
+        )
         args = parser.parse_args()
 
         print("\n🤖 Welcome to AI Robo Advisor")
@@ -90,27 +91,33 @@ def main():
         investment_model = choose_llm_model(all_models, agent="investment_agent")
         portfolio_model = choose_llm_model(all_models, agent="portfolio_agent")
         analyst_model = choose_llm_model(all_models, agent="analyst_agent")
-        
-        investment_llm_agent = get_llm_model(investment_model.provider, investment_model.model_name)
-        portfolio_llm_agent = get_llm_model(portfolio_model.provider, portfolio_model.model_name)
-        analyst_llm_agent = get_llm_model(analyst_model.provider, analyst_model.model_name)
+
+        investment_llm_agent = get_llm_model(
+            investment_model.provider, investment_model.model_name
+        )
+        portfolio_llm_agent = get_llm_model(
+            portfolio_model.provider, portfolio_model.model_name
+        )
+        analyst_llm_agent = get_llm_model(
+            analyst_model.provider, analyst_model.model_name
+        )
 
         # Choose User Preferences
         preferences = get_user_preferences()
 
         # Run the workflow
-        result = run_workflow( 
+        result = run_workflow(
             show_reasoning=args.show_reasoning,
             investment_llm_agent=investment_llm_agent,
             portfolio_llm_agent=portfolio_llm_agent,
             analyst_llm_agent=analyst_llm_agent,  # Fixed: was portfolio_llm_agent
-            preferences=preferences
+            preferences=preferences,
         )
 
         # Display results
-        portfolio = result['data']['portfolio']
-        analysis = result['data']['analysis']['summary']
-        
+        portfolio = result["data"]["portfolio"]
+        analysis = result["data"]["analysis"]["summary"]
+
         print_strategy(portfolio.strategy)
         print_portfolio(portfolio)
         print_analysis_response(analysis)
