@@ -1,16 +1,15 @@
-from graph.state import State
 from data.models import AnalysisAgent
+from graph.state import State
 
 
-def analyze_alignment(state:State) -> State:
-  
+def analyze_alignment(state: State) -> State:
     llm = state["metadata"]["analyst_llm_agent"]
-    portfolio = state["data"]['portfolio']
+    portfolio = state["data"]["portfolio"]
     strategy = portfolio.strategy
 
     # Initialize the analysis dictionary if it doesn't exist
-    if 'analysis' not in state['data']:
-        state['data']['analysis'] = {}
+    if "analysis" not in state["data"]:
+        state["data"]["analysis"] = {}
 
     prompt = f"""
     You are a financial data analyst AI specializing in portfolio alignment analysis. 
@@ -61,9 +60,9 @@ def analyze_alignment(state:State) -> State:
 
     response = llm.with_structured_output(AnalysisAgent).invoke(prompt)
 
-    if state["metadata"]['show_reasoning']:
+    if state["metadata"]["show_reasoning"]:
         print(response.reasoning)
-    
-    state["data"]['analysis']['alignment'] = response
+
+    state["data"]["analysis"]["alignment"] = response
 
     return state
