@@ -1,8 +1,13 @@
-from tools.polygon_api import fetch_histories_concurrently, get_two_year_ago_date, get_today_date
 from data.models import Portfolio
+from tools.polygon_api import (
+    fetch_histories_concurrently,
+    get_today_date,
+    get_two_year_ago_date,
+)
+
 
 def all_data(portfolio: Portfolio, benchmark_ticker: str) -> tuple[dict, dict, dict]:
-    """ Fetch historical data for all portfolio tickers and the benchmark ticker concurrently.
+    """Fetch historical data for all portfolio tickers and the benchmark ticker concurrently.
 
     Args:
         portfolio (Portfolio): The portfolio containing holdings with symbols and weights.
@@ -13,38 +18,40 @@ def all_data(portfolio: Portfolio, benchmark_ticker: str) -> tuple[dict, dict, d
             - portfolio_data: A dictionary where keys are portfolio ticker symbols and values are their historical data.
             - benchmark_data: A dictionary containing the historical data for the benchmark ticker.
             - weights: A dictionary containing the weights of the portfolio tickers.
-    
+
     Raises:
         ValueError: If portfolio is invalid, has no holdings, or data fetching fails.
         TypeError: If portfolio is not a Portfolio instance.
     """
-    
+
     # Validate inputs - check type name to avoid import path issues
-    if not hasattr(portfolio, '__class__') or type(portfolio).__name__ != 'Portfolio':
+    if not hasattr(portfolio, "__class__") or type(portfolio).__name__ != "Portfolio":
         raise TypeError("portfolio must be a Portfolio instance")
-    
+
     if not benchmark_ticker or not isinstance(benchmark_ticker, str):
         raise ValueError("benchmark_ticker must be a non-empty string")
-    
-    if not hasattr(portfolio, 'holdings') or not portfolio.holdings:
+
+    if not hasattr(portfolio, "holdings") or not portfolio.holdings:
         raise ValueError("Portfolio must have holdings")
 
     # Extract tickers and weights from the portfolio
     portfolio_tickers = []
     weights = {}
-    
+
     try:
         for holding in portfolio.holdings:
-            if not hasattr(holding, 'symbol') or not hasattr(holding, 'weight'):
-                raise ValueError("Each holding must have 'symbol' and 'weight' attributes")
-            
+            if not hasattr(holding, "symbol") or not hasattr(holding, "weight"):
+                raise ValueError(
+                    "Each holding must have 'symbol' and 'weight' attributes"
+                )
+
             if not holding.symbol:
                 raise ValueError("Holding symbol cannot be empty")
-            
+
             portfolio_tickers.append(holding.symbol)
             weights[holding.symbol] = holding.weight
     except AttributeError as e:
-        raise ValueError(f"Invalid portfolio holdings structure: {str(e)}")
+        raise ValueError(f"Invalid portfolio holdings structure: {e!s}")
 
     if not portfolio_tickers:
         raise ValueError("Portfolio contains no valid tickers")
@@ -62,9 +69,9 @@ def all_data(portfolio: Portfolio, benchmark_ticker: str) -> tuple[dict, dict, d
         start_date = get_two_year_ago_date()
         end_date = get_today_date()
     except Exception as e:
-        raise ValueError(f"Failed to get date range: {str(e)}")
-    
-    timespan = 'day'
+        raise ValueError(f"Failed to get date range: {e!s}")
+
+    timespan = "day"
 
     # Fetch all data concurrently
     try:
@@ -72,26 +79,34 @@ def all_data(portfolio: Portfolio, benchmark_ticker: str) -> tuple[dict, dict, d
             tickers=all_tickers_to_fetch,
             timespan=timespan,
             start_date=start_date,
-            end_date=end_date
+            end_date=end_date,
         )
     except Exception as e:
-        raise ValueError(f"Failed to fetch historical data: {str(e)}")
+        raise ValueError(f"Failed to fetch historical data: {e!s}")
 
     if not all_histories:
         raise ValueError("No historical data was retrieved")
-    
+
     if benchmark_ticker not in all_histories:
-        raise ValueError(f"Failed to fetch data for benchmark ticker: {benchmark_ticker}")
+        raise ValueError(
+            f"Failed to fetch data for benchmark ticker: {benchmark_ticker}"
+        )
 
     # Check if we have data for all portfolio tickers
-    missing_tickers = [ticker for ticker in portfolio_tickers if ticker not in all_histories]
+    missing_tickers = [
+        ticker for ticker in portfolio_tickers if ticker not in all_histories
+    ]
     if missing_tickers:
-        raise ValueError(f"Failed to fetch data for portfolio tickers: {missing_tickers}")
+        raise ValueError(
+            f"Failed to fetch data for portfolio tickers: {missing_tickers}"
+        )
 
     # Validate that we have sufficient data
     for ticker, data in all_histories.items():
         if not data or len(data) < 2:
-            raise ValueError(f"Insufficient data for ticker {ticker}: need at least 2 data points")
+            raise ValueError(
+                f"Insufficient data for ticker {ticker}: need at least 2 data points"
+            )
 
     # Separate the benchmark data from the portfolio data
     benchmark_data = all_histories.pop(benchmark_ticker)
