@@ -1,5 +1,6 @@
 import pytest
-from src.llm.models import load_models, get_llm_model, ModelProvider
+
+from src.llm.models import ModelProvider, get_llm_model, load_models
 
 PROVIDER_ENV = {
     ModelProvider.ANTHROPIC: "ANTHROPIC_API_KEY",
@@ -12,7 +13,7 @@ PROVIDER_ENV = {
 def test_load_models_includes_all_providers():
     all_models = load_models()
     providers_in_models = {m.provider for m in all_models.models}
-    for prov in PROVIDER_ENV.keys():
+    for prov in PROVIDER_ENV:
         assert prov in providers_in_models, f"{prov} missing from AVAILABLE_MODELS"
 
 
@@ -20,15 +21,15 @@ def test_load_models_includes_all_providers():
 def test_get_llm_model_requires_env_key(provider, env_name, monkeypatch):
     # ensure env not present by patching config to return None
     monkeypatch.delenv(env_name, raising=False)
-    
+
     # Mock the config function to always return None for the specific env_name
     def mock_config(key, default=None):
         if key == env_name:
             return None
         return default
-    
+
     monkeypatch.setattr("src.llm.models.config", mock_config)
-    
+
     with pytest.raises(ValueError) as exc:
         get_llm_model(provider, "dummy-model")
     assert env_name in str(exc.value)
@@ -46,12 +47,14 @@ def test_get_llm_model_calls_init_chat_model(provider, env_name, monkeypatch):
 
     # For GROQ, we need to mock the ChatGroq import to force fallback to init_chat_model
     if provider == ModelProvider.GROQ:
+
         def mock_import(name, *args, **kwargs):
             if name == "langchain_groq":
                 raise ImportError("Mocked import error for testing")
             return __import__(name, *args, **kwargs)
-        
+
         import builtins
+
         original_import = builtins.__import__
         monkeypatch.setattr("builtins.__import__", mock_import)
 
