@@ -1,21 +1,22 @@
 """
 Tests for Polygon API tool integration.
 """
+
+from unittest.mock import Mock, patch
+
 import pytest
-from unittest.mock import Mock, patch, MagicMock
-from datetime import datetime
 
 from src.tools.polygon_api import (
+    POLYGON_AVAILABLE,
+    fetch_histories_concurrently,
     get_stock_history,
     history_to_dict,
-    fetch_histories_concurrently,
-    POLYGON_AVAILABLE
 )
-
 
 # ============================================================================
 # Polygon API Availability Tests
 # ============================================================================
+
 
 @pytest.mark.unit
 class TestPolygonAvailability:
@@ -29,6 +30,7 @@ class TestPolygonAvailability:
 # ============================================================================
 # History to Dict Tests
 # ============================================================================
+
 
 @pytest.mark.unit
 class TestHistoryToDict:
@@ -44,7 +46,7 @@ class TestHistoryToDict:
         mock_item1.low = 99.0
         mock_item1.close = 103.0
         mock_item1.volume = 1000000
-        
+
         mock_item2 = Mock()
         mock_item2.timestamp = 1609545600000  # 2021-01-02 in milliseconds
         mock_item2.open = 103.0
@@ -52,18 +54,18 @@ class TestHistoryToDict:
         mock_item2.low = 102.0
         mock_item2.close = 106.0
         mock_item2.volume = 1200000
-        
+
         history = [mock_item1, mock_item2]
-        
+
         result = history_to_dict(history)
-        
+
         assert len(result) == 2
-        assert all('date' in item for item in result)
-        assert all('open' in item for item in result)
-        assert all('close' in item for item in result)
-        assert all('high' in item for item in result)
-        assert all('low' in item for item in result)
-        assert all('volume' in item for item in result)
+        assert all("date" in item for item in result)
+        assert all("open" in item for item in result)
+        assert all("close" in item for item in result)
+        assert all("high" in item for item in result)
+        assert all("low" in item for item in result)
+        assert all("volume" in item for item in result)
 
     def test_history_to_dict_empty(self):
         """Test converting empty history."""
@@ -79,20 +81,21 @@ class TestHistoryToDict:
         mock_item.low = 99.0
         mock_item.close = 103.0
         mock_item.volume = 1000000
-        
+
         result = history_to_dict([mock_item])
-        
+
         # Check date format (should be 'YYYY-MM-DD HH:MM:SS')
-        assert 'date' in result[0]
-        assert isinstance(result[0]['date'], str)
+        assert "date" in result[0]
+        assert isinstance(result[0]["date"], str)
         # Date should contain year, month, day, and time
-        assert '-' in result[0]['date']
-        assert ':' in result[0]['date']
+        assert "-" in result[0]["date"]
+        assert ":" in result[0]["date"]
 
 
 # ============================================================================
 # Get Stock History Tests
 # ============================================================================
+
 
 @pytest.mark.unit
 @pytest.mark.api
@@ -101,20 +104,20 @@ class TestGetStockHistory:
 
     def test_get_stock_history_polygon_unavailable(self):
         """Test get_stock_history when Polygon is unavailable."""
-        with patch('src.tools.polygon_api.POLYGON_AVAILABLE', False):
-            result = get_stock_history('AAPL', 'day', '2023-01-01', '2023-12-31')
+        with patch("src.tools.polygon_api.POLYGON_AVAILABLE", False):
+            result = get_stock_history("AAPL", "day", "2023-01-01", "2023-12-31")
             assert result == []
 
-    @patch('src.tools.polygon_api.RESTClient')
+    @patch("src.tools.polygon_api.RESTClient")
     def test_get_stock_history_success(self, mock_rest_client):
         """Test successful stock history retrieval."""
         if not POLYGON_AVAILABLE:
             pytest.skip("Polygon not available")
-        
+
         # Mock the client and response
         mock_client_instance = Mock()
         mock_rest_client.return_value = mock_client_instance
-        
+
         mock_agg = Mock()
         mock_agg.timestamp = 1609459200000
         mock_agg.open = 100.0
@@ -122,32 +125,33 @@ class TestGetStockHistory:
         mock_agg.low = 99.0
         mock_agg.close = 103.0
         mock_agg.volume = 1000000
-        
+
         mock_client_instance.get_aggs.return_value = [mock_agg]
-        
-        result = get_stock_history('AAPL', 'day', '2023-01-01', '2023-12-31')
-        
+
+        result = get_stock_history("AAPL", "day", "2023-01-01", "2023-12-31")
+
         assert isinstance(result, list)
         assert len(result) > 0
 
-    @patch('src.tools.polygon_api.RESTClient')
+    @patch("src.tools.polygon_api.RESTClient")
     def test_get_stock_history_api_error(self, mock_rest_client):
         """Test handling of API errors."""
         if not POLYGON_AVAILABLE:
             pytest.skip("Polygon not available")
-        
+
         mock_client_instance = Mock()
         mock_rest_client.return_value = mock_client_instance
         mock_client_instance.get_aggs.side_effect = Exception("API Error")
-        
-        result = get_stock_history('INVALID', 'day', '2023-01-01', '2023-12-31')
-        
+
+        result = get_stock_history("INVALID", "day", "2023-01-01", "2023-12-31")
+
         assert result == []
 
 
 # ============================================================================
 # Fetch Histories Concurrently Tests
 # ============================================================================
+
 
 @pytest.mark.unit
 @pytest.mark.api
@@ -156,86 +160,73 @@ class TestFetchHistoriesConcurrently:
 
     def test_fetch_histories_polygon_unavailable(self):
         """Test fetch_histories_concurrently when Polygon is unavailable."""
-        with patch('src.tools.polygon_api.POLYGON_AVAILABLE', False):
+        with patch("src.tools.polygon_api.POLYGON_AVAILABLE", False):
             result = fetch_histories_concurrently(
-                ['AAPL', 'GOOGL'],
-                'day',
-                '2023-01-01',
-                '2023-12-31'
+                ["AAPL", "GOOGL"], "day", "2023-01-01", "2023-12-31"
             )
             assert result == {}
 
-    @patch('src.tools.polygon_api.get_stock_history')
+    @patch("src.tools.polygon_api.get_stock_history")
     def test_fetch_histories_multiple_tickers(self, mock_get_history):
         """Test fetching histories for multiple tickers."""
         if not POLYGON_AVAILABLE:
             pytest.skip("Polygon not available")
-        
+
         # Mock return values for different tickers
         def mock_history_side_effect(symbol, *args):
-            if symbol == 'AAPL':
-                return [{'date': '2023-01-01', 'close': 100}]
-            elif symbol == 'GOOGL':
-                return [{'date': '2023-01-01', 'close': 200}]
+            if symbol == "AAPL":
+                return [{"date": "2023-01-01", "close": 100}]
+            elif symbol == "GOOGL":
+                return [{"date": "2023-01-01", "close": 200}]
             return []
-        
+
         mock_get_history.side_effect = mock_history_side_effect
-        
+
         result = fetch_histories_concurrently(
-            ['AAPL', 'GOOGL'],
-            'day',
-            '2023-01-01',
-            '2023-12-31'
+            ["AAPL", "GOOGL"], "day", "2023-01-01", "2023-12-31"
         )
-        
-        assert 'AAPL' in result
-        assert 'GOOGL' in result
+
+        assert "AAPL" in result
+        assert "GOOGL" in result
         assert len(result) == 2
 
-    @patch('src.tools.polygon_api.get_stock_history')
+    @patch("src.tools.polygon_api.get_stock_history")
     def test_fetch_histories_empty_tickers(self, mock_get_history):
         """Test fetching histories with empty ticker list."""
         if not POLYGON_AVAILABLE:
             pytest.skip("Polygon not available")
-        
-        result = fetch_histories_concurrently(
-            [],
-            'day',
-            '2023-01-01',
-            '2023-12-31'
-        )
-        
+
+        result = fetch_histories_concurrently([], "day", "2023-01-01", "2023-12-31")
+
         assert result == {}
 
-    @patch('src.tools.polygon_api.get_stock_history')
+    @patch("src.tools.polygon_api.get_stock_history")
     def test_fetch_histories_error_handling(self, mock_get_history):
         """Test that errors in individual fetches don't break everything."""
         if not POLYGON_AVAILABLE:
             pytest.skip("Polygon not available")
-        
+
         def mock_history_side_effect(symbol, *args):
-            if symbol == 'AAPL':
-                return [{'date': '2023-01-01', 'close': 100}]
-            elif symbol == 'ERROR':
+            if symbol == "AAPL":
+                return [{"date": "2023-01-01", "close": 100}]
+            elif symbol == "ERROR":
                 raise Exception("API Error")
             return []
-        
+
         mock_get_history.side_effect = mock_history_side_effect
-        
+
         result = fetch_histories_concurrently(
-            ['AAPL', 'ERROR', 'GOOGL'],
-            'day',
-            '2023-01-01',
-            '2023-12-31'
+            ["AAPL", "ERROR", "GOOGL"], "day", "2023-01-01", "2023-12-31"
         )
-        
+
         # Should still get AAPL data even though ERROR failed
-        assert 'AAPL' in result or isinstance(result, dict)
+        assert "AAPL" in result or isinstance(result, dict)
 
 
 # ============================================================================
 # Integration Tests
 # ============================================================================
+
 
 @pytest.mark.integration
 @pytest.mark.api
